+++ PYTHON/vehiclemodels/utils/longitudinal_parameters.py
@@ -1,7 +1,16 @@
+from dataclasses import dataclass
+from typing import Optional
+
+
+@dataclass
 class LongitudinalParameters:
-    def __init__(self):
-        # constraints regarding longitudinal dynamics
-        self.v_min = None  # minimum velocity [m/s]
-        self.v_max = None  # minimum velocity [m/s]
-        self.v_switch = None  # switching velocity [m/s]
-        self.a_max = None  # maximum absolute acceleration [m/s^2]
+    """
+    Class defines all parameters related to the longitudinal dynamics
+    """
+    # constraints regarding longitudinal dynamics
+    v_min: Optional[float] = None  # minimum velocity [m/s]
+    v_max: Optional[float] = None  # minimum velocity [m/s]
+    v_switch: Optional[float] = None  # switching velocity [m/s]
+    a_max: Optional[float] = None  # maximum absolute acceleration [m/s^2]
+    j_max: Optional[float] = None  # maximum longitudinal jerk [m/s^3]
+    j_dot_max: Optional[float] = None  # maximum change of longitudinal jerk [m/s^4]