+++ PYTHON/vehiclemodels/utils/steering_constraints.py
@@ -1,33 +1,20 @@
 def steering_constraints(steering_angle, steering_velocity, p):
-    # steeringConstraints - adjusts the steering velocity based on steering
-    # constraints
-    #
-    # Syntax:  
-    #    steeringConstraints(steering_angle,steering_velocity,p)
-    #
-    # Inputs:
-    #    steering_angle - steering angle
-    #    steering_velocity - steering velocity
-    #    p - steering parameter structure
-    #
-    # Outputs:
-    #    steering_velocity - steering velocity
-    #
-    # Example: 
-    #
-    # Other m-files required: none
-    # Subfunctions: none
-    # MAT-files required: none
-    #
-    # See also: ---
+    """
+    steering_constraints - adjusts the steering velocity based on steering
 
-    # Author:       Matthias Althoff
-    # Written:      15-December-2017
-    # Last update:  ---
-    # Last revision:---
+    Inputs:
+        :param steering_angle - steering angle
+        :param steering_velocity - steering velocity
+        :params p - steering parameter structure
 
-    # ------------- BEGIN CODE --------------
+    Outputs:
+        :return steering_velocity - steering velocity
 
+    Author: Matthias Althoff
+    Written: 15-December-2017
+    Last update: ---
+    Last revision: ---
+    """
     # steering limit reached?
     if (steering_angle <= p.min and steering_velocity <= 0) or (steering_angle >= p.max and steering_velocity >= 0):
         steering_velocity = 0
@@ -38,4 +25,17 @@ def steering_constraints(steering_angle, steering_velocity, p):
 
     return steering_velocity
 
-    # ------------- END OF CODE --------------
+
+def kappa_dot_dot_constraints(kappa_dot_dot, kappa_dot, p):
+    """
+    input constraints for kappa_dot_dot: adjusts kappa_dot_dot if kappa_dot limit (i.e., maximum curvature rate)
+    or input bounds are reached
+    """
+    if (kappa_dot < -p.kappa_dot_max and kappa_dot_dot < 0.) \
+            or (kappa_dot > p.kappa_dot_max and kappa_dot_dot > 0.):
+        # kappa_dot limit reached
+        kappa_dot_dot = 0.
+    elif abs(kappa_dot_dot) >= p.kappa_dot_dot_max:
+        # input bounds reached
+        kappa_dot_dot = p.kappa_dot_dot_max
+    return kappa_dot_dot