+++ PYTHON/vehiclemodels/utils/steering_parameters.py
@@ -1,7 +1,16 @@
-class SteeringParameters():
-    def __init__(self):
-        #constraints regarding steering
-        self.min = []  #minimum steering angle [rad]
-        self.max = []  #maximum steering angle [rad]
-        self.v_min = []  #minimum steering velocity [rad/s]
-        self.v_max = []  #maximum steering velocity [rad/s]
+from dataclasses import dataclass
+from typing import Optional
+
+
+@dataclass
+class SteeringParameters:
+    """
+    Class defines all steering related parameters
+    """
+    # constraints regarding steering
+    min: Optional[float] = None  # minimum steering angle [rad]
+    max: Optional[float] = None  # maximum steering angle [rad]
+    v_min: Optional[float] = None  # minimum steering velocity [rad/s]
+    v_max: Optional[float] = None  # maximum steering velocity [rad/s]
+    kappa_dot_max: Optional[float] = None  # maximum curvature rate
+    kappa_dot_dot_max: Optional[float] = None  # maximum curvature rate rate