+++ PYTHON/vehiclemodels/utils/tireParameters.py
@@ -1,38 +1,45 @@
-class TireParameters():
-    def __init__(self):
-        # tire parameters from ADAMS handbook
-        # longitudinal coefficients
-        self.p_cx1 = None  # Shape factor Cfx for longitudinal force
-        self.p_dx1 = None  # Longitudinal friction Mux at Fznom
-        self.p_dx3 = None  # Variation of friction Mux with camber
-        self.p_ex1 = None  # Longitudinal curvature Efx at Fznom
-        self.p_kx1 = None  # Longitudinal slip stiffness Kfx/Fz at Fznom
-        self.p_hx1 = None  # Horizontal shift Shx at Fznom
-        self.p_vx1 = None  # Vertical shift Svx/Fz at Fznom
-        self.r_bx1 = None  # Slope factor for combined slip Fx reduction
-        self.r_bx2 = None  # Variation of slope Fx reduction with kappa
-        self.r_cx1 = None  # Shape factor for combined slip Fx reduction
-        self.r_ex1 = None  # Curvature factor of combined Fx
-        self.r_hx1 = None  # Shift factor for combined slip Fx reduction
+from dataclasses import dataclass
+from typing import Optional
 
-        # lateral coefficients
-        self.p_cy1 = None  # Shape factor Cfy for lateral forces
-        self.p_dy1 = None  # Lateral friction Muy
-        self.p_dy3 = None  # Variation of friction Muy with squared camber
-        self.p_ey1 = None  # Lateral curvature Efy at Fznom
-        self.p_ky1 = None  # Maximum value of stiffness Kfy/Fznom
-        self.p_hy1 = None  # Horizontal shift Shy at Fznom
-        self.p_hy3 = None  # Variation of shift Shy with camber
-        self.p_vy1 = None  # Vertical shift in Svy/Fz at Fznom
-        self.p_vy3 = None  # Variation of shift Svy/Fz with camber
-        self.r_by1 = None  # Slope factor for combined Fy reduction
-        self.r_by2 = None  # Variation of slope Fy reduction with alpha
-        self.r_by3 = None  # Shift term for alpha in slope Fy reduction
-        self.r_cy1 = None  # Shape factor for combined Fy reduction
-        self.r_ey1 = None  # Curvature factor of combined Fy
-        self.r_hy1 = None  # Shift factor for combined Fy reduction
-        self.r_vy1 = None  # Kappa induced side force Svyk/Muy*Fz at Fznom
-        self.r_vy3 = None  # Variation of Svyk/Muy*Fz with camber
-        self.r_vy4 = None  # Variation of Svyk/Muy*Fz with alpha
-        self.r_vy5 = None  # Variation of Svyk/Muy*Fz with kappa
-        self.r_vy6 = None  # Variation of Svyk/Muy*Fz with atan(kappa)
+
+@dataclass
+class TireParameters:
+    """
+    Class defines all Tire Parameters
+    """
+    # tire parameters from ADAMS handbook
+    # longitudinal coefficients
+    p_cx1: Optional[float] = None  # Shape factor Cfx for longitudinal force
+    p_dx1: Optional[float] = None  # Longitudinal friction Mux at Fznom
+    p_dx3: Optional[float] = None  # Variation of friction Mux with camber
+    p_ex1: Optional[float] = None  # Longitudinal curvature Efx at Fznom
+    p_kx1: Optional[float] = None  # Longitudinal slip stiffness Kfx/Fz at Fznom
+    p_hx1: Optional[float] = None  # Horizontal shift Shx at Fznom
+    p_vx1: Optional[float] = None  # Vertical shift Svx/Fz at Fznom
+    r_bx1: Optional[float] = None  # Slope factor for combined slip Fx reduction
+    r_bx2: Optional[float] = None  # Variation of slope Fx reduction with kappa
+    r_cx1: Optional[float] = None  # Shape factor for combined slip Fx reduction
+    r_ex1: Optional[float] = None  # Curvature factor of combined Fx
+    r_hx1: Optional[float] = None  # Shift factor for combined slip Fx reduction
+
+    # lateral coefficients
+    p_cy1: Optional[float] = None  # Shape factor Cfy for lateral forces
+    p_dy1: Optional[float] = None  # Lateral friction Muy
+    p_dy3: Optional[float] = None  # Variation of friction Muy with squared camber
+    p_ey1: Optional[float] = None  # Lateral curvature Efy at Fznom
+    p_ky1: Optional[float] = None  # Maximum value of stiffness Kfy/Fznom
+    p_hy1: Optional[float] = None  # Horizontal shift Shy at Fznom
+    p_hy3: Optional[float] = None  # Variation of shift Shy with camber
+    p_vy1: Optional[float] = None  # Vertical shift in Svy/Fz at Fznom
+    p_vy3: Optional[float] = None  # Variation of shift Svy/Fz with camber
+    r_by1: Optional[float] = None  # Slope factor for combined Fy reduction
+    r_by2: Optional[float] = None  # Variation of slope Fy reduction with alpha
+    r_by3: Optional[float] = None  # Shift term for alpha in slope Fy reduction
+    r_cy1: Optional[float] = None  # Shape factor for combined Fy reduction
+    r_ey1: Optional[float] = None  # Curvature factor of combined Fy
+    r_hy1: Optional[float] = None  # Shift factor for combined Fy reduction
+    r_vy1: Optional[float] = None  # Kappa induced side force Svyk/Muy*Fz at Fznom
+    r_vy3: Optional[float] = None  # Variation of Svyk/Muy*Fz with camber
+    r_vy4: Optional[float] = None  # Variation of Svyk/Muy*Fz with alpha
+    r_vy5: Optional[float] = None  # Variation of Svyk/Muy*Fz with kappa
+    r_vy6: Optional[float] = None  # Variation of Svyk/Muy*Fz with atan(kappa)