+++ PYTHON/vehiclemodels/vehicle_dynamics_ks.py
@@ -34,35 +34,32 @@ def vehicle_dynamics_ks(x, u_init, p):
     Last update: 16-December-2017
     Last revision: ---
     """
+    # create equivalent kinematic single-track parameters
+    l = p.a + p.b
 
-    #------------- BEGIN CODE --------------
+    # states
+    # x1 = x-position in a global coordinate system
+    # x2 = y-position in a global coordinate system
+    # x3 = steering angle of front wheels
+    # x4 = velocity in x-direction
+    # x5 = yaw angle
 
-    #create equivalent kinematic single-track parameters
-    l = p.a + p.b 
+    # inputs
+    # u1 = steering angle velocity of front wheels
+    # u2 = longitudinal acceleration
 
-    #states
-    #x1 = x-position in a global coordinate system
-    #x2 = y-position in a global coordinate system
-    #x3 = steering angle of front wheels
-    #x4 = velocity in x-direction
-    #x5 = yaw angle
+    # consider steering constraints
+    u = list()
+    u.append(steering_constraints(x[2], u_init[0], p.steering))  # different name u_init/u due to side effects of u
+    # consider acceleration constraints
+    u.append(
+        acceleration_constraints(x[3], u_init[1], p.longitudinal))  # different name u_init/u due to side effects of u
 
-    #u1 = steering angle velocity of front wheels
-    #u2 = longitudinal acceleration
-    
-    #consider steering constraints
-    u = [];
-    u.append(steering_constraints(x[2], u_init[0], p.steering)) # different name u_init/u due to side effects of u
-    #consider acceleration constraints
-    u.append(acceleration_constraints(x[3], u_init[1], p.longitudinal)) # different name u_init/u due to side effects of u
+    # system dynamics
+    f = [x[3] * math.cos(x[4]),
+         x[3] * math.sin(x[4]),
+         u[0],
+         u[1],
+         x[3] / l * math.tan(x[2])]
 
-    #system dynamics
-    f = [x[3]*math.cos(x[4]), 
-        x[3]*math.sin(x[4]), 
-        u[0], 
-        u[1], 
-        x[3]/l*math.tan(x[2])]
-    
     return f
-
-    #------------- END OF CODE --------------