+++ PYTHON/vehiclemodels/vehicle_dynamics_std.py
@@ -114,9 +114,11 @@ def vehicle_dynamics_std(x, u_init, p):
         T_E = 0.
 
     # system dynamics
-    d_v = 1 / m * (-F_yf * math.sin(x[2] - x[6]) + F_yr * math.sin(x[6]) + F_xr * math.cos(x[6]) + F_xf * math.cos(x[2] - x[6]))
+    d_v = 1 / m * (-F_yf * math.sin(x[2] - x[6]) + F_yr * math.sin(x[6]) + F_xr * math.cos(x[6]) +
+                   F_xf * math.cos(x[2] - x[6]))
     dd_psi = 1 / I * (F_yf * math.cos(x[2]) * lf - F_yr * lr + F_xf * math.sin(x[2]) * lf)
-    d_beta = -x[5] + 1 / (m * x[3]) * (F_yf * math.cos(x[2] - x[6]) + F_yr * math.cos(x[6]) - F_xr * math.sin(x[6]) + F_xf * math.sin(x[2] - x[6])) if x[3] > v_min else 0
+    d_beta = -x[5] + 1 / (m * x[3]) * (F_yf * math.cos(x[2] - x[6]) + F_yr * math.cos(x[6]) - F_xr * math.sin(x[6]) +
+                                       F_xf * math.sin(x[2] - x[6])) if x[3] > v_min else 0
 
     # wheel dynamics (negative wheel spin forbidden)
     d_omega_f = 1 / p.I_y_w * (-p.R_w * F_xf + p.T_sb * T_B + p.T_se * T_E) if x[7] >= 0 else 0
@@ -125,6 +127,8 @@ def vehicle_dynamics_std(x, u_init, p):
     x[8] = max(0, x[8])
 
     # *** Mix with kinematic model at low speeds ***
+    # Due to errors when using the scipy.odeint with a "hard" switch to the kinematic model, we overblend both models
+    # around the switching velocity to achieve a "smoother" transition between both models.
     # kinematic system dynamics
     x_ks = [x[0], x[1], x[2], x[3], x[4]]
     f_ks = vehicle_dynamics_ks_cog(x_ks, u, p)
@@ -153,122 +157,3 @@ def vehicle_dynamics_std(x, u_init, p):
          w_std * d_omega_r + w_ks * d_omega_r_ks]
 
     return f
-
-
-# ***********************************
-# Test Model
-# ***********************************
-if __name__ == "__main__":
-    from parameters_vehicle2 import parameters_vehicle2
-    from init_std import init_std
-    from scipy.integrate import odeint, solve_ivp
-    import numpy
-    import matplotlib.pyplot as plt
-
-    def func_STD(x, t, u, p):
-        f = vehicle_dynamics_std(x, u, p)
-        return f
-
-    def func_STD_2(t, x, u, p):
-        f = vehicle_dynamics_std(x, u, p)
-        return f
-
-    def plot_odeint(x_std, t):
-        plt.plot([tmp[0] for tmp in x_std], [tmp[1] for tmp in x_std])
-        plt.title('position')
-        plt.xlabel('x')
-        plt.ylabel('y')
-        plt.show()
-
-        plt.plot(t, [tmp[2] for tmp in x_std])
-        plt.title('steering angle')
-        plt.show()
-
-        plt.plot(t, [tmp[3] for tmp in x_std])
-        plt.title('velocity')
-        plt.show()
-
-        plt.plot(t, [tmp[4] for tmp in x_std])
-        plt.title('yaw angle')
-        plt.show()
-
-        plt.plot(t, [tmp[5] for tmp in x_std])
-        plt.title('yaw rate')
-        plt.show()
-
-        plt.plot(t, [tmp[6] for tmp in x_std])
-        plt.title('slip angle')
-        plt.show()
-
-        plt.plot(t, [tmp[7] for tmp in x_std])
-        plt.title('omega front')
-        plt.show()
-
-        plt.plot(t, [tmp[8] for tmp in x_std])
-        plt.title('omega rear')
-        plt.show()
-
-    def plot_solve_ivp(x_std):
-        plt.plot(x_std.y[0], x_std.y[1])
-        plt.title('position')
-        plt.xlabel('x')
-        plt.ylabel('y')
-        plt.show()
-
-        plt.plot(x_std.t, x_std.y[2])
-        plt.title('steering angle')
-        plt.show()
-
-        plt.plot(x_std.t, x_std.y[3])
-        plt.title('velocity')
-        plt.show()
-
-        plt.plot(x_std.t, x_std.y[4])
-        plt.title('yaw angle')
-        plt.show()
-
-        plt.plot(x_std.t, x_std.y[5])
-        plt.title('yaw rate')
-        plt.show()
-
-        plt.plot(x_std.t, x_std.y[6])
-        plt.title('slip angle')
-        plt.show()
-
-        plt.plot(x_std.t, x_std.y[7])
-        plt.title('omega front')
-        plt.show()
-
-        plt.plot(x_std.t, x_std.y[8])
-        plt.title('omega rear')
-        plt.show()
-
-    # load parameters
-    p = parameters_vehicle2()
-    g = 9.81  # [m/s^2]
-
-    # set options --------------------------------------------------------------
-    tStart = 0  # start time
-    tFinal = 1  # start time
-
-    delta0 = 0.0
-    vel0 = 0
-    Psi0 = 0
-    dotPsi0 = 0
-    beta0 = 0
-    sy0 = 0
-    initialState = [0, sy0, delta0, vel0, Psi0, dotPsi0, beta0]  # initial state for simulation
-    x0_STD = init_std(initialState, p)  # initial state for multi-body model
-    # --------------------------------------------------------------------------
-
-    t = numpy.arange(0, tFinal, 0.01)
-    u = [0.0, 0.6 * g]
-
-    # using odeint
-    # x_std, out_dict = odeint(func_STD, x0_STD, t, args=(u, p), printmessg=True, full_output=True)
-    # print(out_dict)
-    # plot_odeint(x_std, t)
-
-    # using solve_ivp
-    x_std = solve_ivp(func_STD_2, (tStart, tFinal), x0_STD, args=(u, p))
-    plot_solve_ivp(x_std)